FLIX-198: Customer can not create a requets if previous request in status PAYMENT_WAITING

+++ src/main/java/com/fu/flix/configuration/AppConf.java
@@ -21,7 +21,7 @@ public class AppConf {
     private Double vat;
     private Double profitRate;
     private Long minTimeFined;
-    private Double fine;
+    private Long fine;
     private VnPayInfo vnPayInfo;
 
     @Data
